refactor transcribe add route
