+++ .moai/specs/SPEC-PORTAL-TRANSCRIBE-ADD-CLEANUP-001/progress.md
@@ -1,0 +1,9 @@
+## SPEC-PORTAL-TRANSCRIBE-ADD-CLEANUP-001 Progress
+
+- Started: 2026-05-13T09:05:01+02:00
+- Phase 0.5 complete: memory_guard=disabled, strategy=full
+- Phase 0.9 complete: detected_language_skills=typescript
+- Phase 0.95 complete: scale-based mode=focused (frontend route refactor)
+- Phase 1 complete: analyzed `transcribe/add.tsx`; actual modes are record and upload, with no paste URL mode.
+- Phase 2 complete: extracted route helpers, mutation hooks, recording form, upload form, and result cards.
+- Phase 2.75 complete: targeted unit test, ESLint, TypeScript build, and production build passed.
